Apply suggestions
