+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs
@@ -77,8 +77,8 @@ where
 
 import Cardano.Crypto.Hash.Class (HashAlgorithm)
 import Cardano.Ledger.Address (Addr (..), RewardAcnt (..))
-import Cardano.Ledger.Allegra.Tx (validateTimelock)
 import Cardano.Ledger.Allegra.Core ()
+import Cardano.Ledger.Allegra.Tx (validateTimelock)
 import Cardano.Ledger.Alonzo.Data (Data, hashData)
 import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.Alonzo.PParams
