+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/STSTestUtils.hs
@@ -64,7 +64,6 @@ import Cardano.Ledger.BaseTypes
     mkTxIxPartial,
   )
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Conway.Rules ()
 import Cardano.Ledger.Core hiding (TranslationError)
 import Cardano.Ledger.Credential
   ( Credential (..),
@@ -324,7 +323,7 @@ testBBODY wit@(BBODY proof) initialSt block expected pparams =
         Alonzo _ -> runSTS wit (TRC (env, initialSt, block)) (genericCont "" expected)
         Babbage _ -> runSTS wit (TRC (env, initialSt, block)) (genericCont "" expected)
         -- TODO re-enable this once we have added all the new rules to Conway
-        --Conway _ -> runSTS wit (TRC (env, initialSt, block)) (genericCont "" expected)
+        -- Conway _ -> runSTS wit (TRC (env, initialSt, block)) (genericCont "" expected)
         Conway _ -> undefined
         other -> error ("We cannot testBBODY in era " ++ show other)
 
@@ -379,7 +378,7 @@ testUTXOspecialCase wit@(UTXOW proof) utxo pparam tx expected =
         Alonzo _ -> runSTS wit (TRC (env, state, tx)) (specialCont proof expected)
         Babbage _ -> runSTS wit (TRC (env, state, tx)) (specialCont proof expected)
         -- TODO re-enable this once we have added all the new rules to Conway
-        --Conway _ -> runSTS wit (TRC (env, state, tx)) (specialCont proof expected)
+        -- Conway _ -> runSTS wit (TRC (env, state, tx)) (specialCont proof expected)
         other -> error ("Cannot use specialCase in era " ++ show other)
 
 -- ======================================================================
@@ -408,7 +407,7 @@ testUTXOWwith wit@(UTXOW proof) cont utxo pparams tx expected =
       state = smartUTxOState utxo (Coin 0) (Coin 0) def
    in case proof of
         -- TODO re-enable this once we have added all the new rules to Conway
-        --Conway _ -> runSTS wit (TRC (env, state, tx)) (cont expected)
+        -- Conway _ -> runSTS wit (TRC (env, state, tx)) (cont expected)
         Conway _ -> undefined
         Babbage _ -> runSTS wit (TRC (env, state, tx)) (cont expected)
         Alonzo _ -> runSTS wit (TRC (env, state, tx)) (cont expected)
