+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Fields.hs
@@ -59,8 +59,8 @@ import Cardano.Ledger.BaseTypes
   )
 import Cardano.Ledger.Binary (sizedValue)
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Conway.Core (GovernanceActionInfo (..), Vote)
-import Cardano.Ledger.Conway.Delegation.Certificates (transDCert, ConwayDCert)
+import Cardano.Ledger.Conway.Governance (GovernanceActionInfo (..), Vote)
+import Cardano.Ledger.Conway.Delegation.Certificates (ConwayDCert, transDCert)
 import Cardano.Ledger.Conway.TxBody (ConwayTxBody (..))
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..), StakeReference (..))
