+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Properties.hs
@@ -16,7 +16,6 @@ import Cardano.Ledger.Alonzo.Tx (AlonzoTxBody (..), IsValid (..))
 import qualified Cardano.Ledger.Babbage.PParams (BabbagePParamsHKD (..))
 import Cardano.Ledger.Babbage.TxBody (BabbageTxBody (..))
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Conway.Rules ()
 import Cardano.Ledger.Core
 import Cardano.Ledger.Keys (GenDelegs (..))
 import Cardano.Ledger.Pretty (PrettyA (..), ppList)
@@ -278,9 +277,8 @@ incrementalStake :: GenSize -> TestTree
 incrementalStake genSize =
   testGroup
     "Incremental Stake invariant holds"
-    [ 
-      -- TODO re-enable this once we have added all the new rules to Conway
-      --incrementStakeInvariant (Conway Mock) genSize,
+    [ -- TODO re-enable this once we have added all the new rules to Conway
+      -- incrementStakeInvariant (Conway Mock) genSize,
       incrementStakeInvariant (Babbage Mock) genSize,
       incrementStakeInvariant (Alonzo Mock) genSize,
       incrementStakeInvariant (Mary Mock) genSize,
