+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Same.hs
@@ -25,12 +25,10 @@ import Cardano.Ledger.Alonzo.TxSeq (AlonzoTxSeq (..))
 import Cardano.Ledger.Alonzo.TxWits (AlonzoTxWits (..), Redeemers (..), TxDats (..))
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
 import Cardano.Ledger.Babbage.PParams (BabbagePParamsHKD)
-import Cardano.Ledger.Babbage.Rules ()
 import Cardano.Ledger.Babbage.TxBody (BabbageEraTxBody, BabbageTxBody (..), BabbageTxOut (..))
 import Cardano.Ledger.Binary (sizedValue)
 import Cardano.Ledger.Block (Block (..))
 import Cardano.Ledger.Conway.Core (ConwayEraTxBody)
-import Cardano.Ledger.Conway.Rules ()
 import Cardano.Ledger.Conway.TxBody (ConwayTxBody (..))
 import Cardano.Ledger.Core (EraTxOut (..))
 import qualified Cardano.Ledger.Core as Core
